feat(Sales Invoice): make validation configurable (IBT-86)
Added support for configurable validation behavior with three modes:
- Silen validation : validations run, but no messages
- No validation: disables all validations
- Warning: validations run, but failures only trigger warnings
- Error: validations run, and failures block further actions

This allows more flexible handling of validation logic during form submission or data insertion.

In case the user haven't set a value in E Invoice Settings, the system gave the user a warning that the default value (Silent Validation) is used.

resolves: #159